Refactor commands to their own module
